backend/requirements.txt
flask
pdfplumber
spacy
fuzzywuzzy
scikit-learn
mysql-connector-python
gunicorn
